move all the remaining tests into correct modules
