+++ tests/FilePermissions/PermissionParserTest.php
